[COMMON] soc: samsung: cpupm: fix prevent defect 280570

+++ drivers/soc/samsung/exynos-cpupm.c
@@ -420,7 +420,7 @@ static int cpus_busy(int target_residency, const struct cpumask *cpus)
 		 */
 		moce_ratio = exynos_moce_get_ratio(cpu);
 
-		if (get_sleep_length(cpu) < (target_residency * moce_ratio) / 100)
+		if (get_sleep_length(cpu) < (target_residency * (s64)moce_ratio) / 100)
 			return -EBUSY;
 	}
